Renamed files to lowercase
